django_service/requirements.txt
asgiref==3.8.1
Django==5.1.5
django-leaflet==0.31.0
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
